+++ Helper.h
@@ -1,18 +1,17 @@
 #pragma once
-#include "AEEngine.h"
 struct CooldownTimer
 {
-    f32 duration = 0.0;
-    f32 elapsed = 0.0;
+    double duration = 0.0;
+    double elapsed = 0.0;
     bool active = false;
 
-    void Start(f32 d);
+    void Start(double d);
 
     void Stop();
 
-    bool Update(f32 dt);
+    bool Update(double dt);
 
-    f32 GetProgress() const;
+    double GetProgress() const;
 
     bool IsActive() const;
 };
